Almost enabled sprites to span several grid cells. Required significant thought for the design so that any special SpriteObject parameter can be kept track of with an annotation and properly dispalyed. Some params are locked to any changes. Some params allow changes. By putting a changeListener on the user changes to the width and height attributes, this chanegs the Sprites width and height properties. On chnge for those consumes a consumer. This consumer can be set when the sprite is put into the grid originally. Note that the grid is now composed of new classes that encapsulate both the stackpane of a grid cell, and special logic to create a consumer, keep track of their layer, etc.

+++ src/authoring_UI/SpriteDataConverter.java
@@ -36,8 +36,8 @@ public class SpriteDataConverter {
 	ArrayList<SpriteDataConverter> inventory;
 	Integer[] gridPos;
 	String name;
-	Double width;
-	Double height;
+	Integer width;
+	Integer height;
 	String UUID;
 	String imageURL;
 	String mySavePath;
